Trabajo 2/requirements.txt
appnope==0.1.3
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.38.0
ipykernel==6.21.2
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
wcwidth==0.2.6
